Index fromPairs object by known keys when possible
If the pair values are all of a homogeneous type then we return a record
with values of that type, indexed by the known keys. If the values are
of disparate types then we merge the pairs into an object one by one,
so that if a key appears more than once the rightmost value type is used
in the result.